Rename "Middlewares" namespace to "Middleware" (#2499)
