+++ docs/basic-usage/passport.md
@@ -39,7 +39,7 @@ The extended Client should either provide a `$guard_name` property or a `guardNa
 They should return a string that matches the [configured](https://laravel.com/docs/master/passport#installation) guard name for the passport driver.
 
 ## Middleware
-All middlewares provided by this package work with the Client.
+All middleware provided by this package work with the Client.
 
 Do make sure that you only wrap your routes in the [`client`](https://laravel.com/docs/master/passport#via-middleware) middleware and not the `auth:api` middleware as well.
 Wrapping routes in the `auth:api` middleware currently does not work for the Client Credentials Grant.
