+++ tests/PermissionMiddlewareTest.php
@@ -11,7 +11,7 @@ use InvalidArgumentException;
 use Laravel\Passport\Passport;
 use Spatie\Permission\Contracts\Permission;
 use Spatie\Permission\Exceptions\UnauthorizedException;
-use Spatie\Permission\Middlewares\PermissionMiddleware;
+use Spatie\Permission\Middleware\PermissionMiddleware;
 use Spatie\Permission\Tests\TestModels\UserWithoutHasRoles;
 
 class PermissionMiddlewareTest extends TestCase
@@ -399,15 +399,15 @@ class PermissionMiddlewareTest extends TestCase
     public function the_middleware_can_be_created_with_static_using_method()
     {
         $this->assertSame(
-            'Spatie\Permission\Middlewares\PermissionMiddleware:edit-articles',
+            'Spatie\Permission\Middleware\PermissionMiddleware:edit-articles',
             PermissionMiddleware::using('edit-articles')
         );
         $this->assertEquals(
-            'Spatie\Permission\Middlewares\PermissionMiddleware:edit-articles,my-guard',
+            'Spatie\Permission\Middleware\PermissionMiddleware:edit-articles,my-guard',
             PermissionMiddleware::using('edit-articles', 'my-guard')
         );
         $this->assertEquals(
-            'Spatie\Permission\Middlewares\PermissionMiddleware:edit-articles|edit-news',
+            'Spatie\Permission\Middleware\PermissionMiddleware:edit-articles|edit-news',
             PermissionMiddleware::using(['edit-articles', 'edit-news'])
         );
     }
